add granular access middleware
