+++ src/server/__tests__/testConfigFiles/test-invalid-index-scoped-tier-access.json
@@ -6,6 +6,11 @@
             "tier_access_level": "regular"
         },
         {
+            "index": "subject_granular",
+            "type": "subject",
+            "tier_access_level": "granular"
+        },
+        {
             "index": "subject_private",
             "type": "subject_private",
             "tier_access_level": "private____typo"
